requirements-no-torch.txt
numpy==1.26.3
scipy==1.11.4
joblib==1.2.0
# Cython==0.29.10
pandas==2.1.4
scikit-learn==1.4.0
wfdb==3.2.0
xgboost==2.0.3
opencv-python
torchsummary
tensorboardX
tqdm
gdown
biosppy
deprecated
torch-optimizer
IPython
pyEDFlib
PyWavelets
xmltodict
einops
matplotlib
seaborn
requests
accelerate
nitime
PyYAML
diff-binom-confint
